# backend/spring-boot/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: bugzkit
  description: A production-ready web application template
  termsOfService: https://swagger.io/terms/
  contact:
    email: office@bugzkit.com
  license:
    name: MIT License
    url: https://choosealicense.com/licenses/mit/
  version: 1.0.0
servers:
  - url: http://localhost:8080
  - url: https://localhost:8080
...
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: User registration
      requestBody:
        $ref: "#/components/requestBodies/RegisterUserRequest"
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
              example:
                id: 5
                username: john.doe
                email: john.doe@localhost
                createdAt: 2023-05-18T00:51:50.758738
        400:
          $ref: "#/components/responses/BadRequest"
        409:
          $ref: "#/components/responses/Conflict"
  /auth/tokens:
    post:
      tags:
        - auth
      summary: Create access and refresh tokens
      requestBody:
        $ref: "#/components/requestBodies/AuthTokenRequest"
      responses:
        204:
          description: Auth tokens created successfully
          headers:
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "accessToken=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJVU0VSIl0sImlzc3VlZEF0IjoiMjAyMS0xMC0yMFQxNDozMzo0NC43MDcyOTMzMDlaIiwiZXhwIjoxNjM0NzQxMzI0LCJ1c2VySWQiOjJ9.uXOVA1q-o2DtHmwBAzEfqEm8GLpAhXrYo0rlZ_6NFbBGILhkV74x-Iu9W2uSfSlwp1IfKPCHlR6zWVPvAbhWVw; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=9000"
                  - "refreshToken=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJVU0VSIl0sImlzc3VlZEF0IjoiMjAyMS0xMC0xOVQxMzowMDowMy45MDYzNDQyMDJaIiwiZXhwIjoxNjM1MjUzMjAzLCJ1c2VySWQiOjJ9.RHzh6qyGJEKYdvCuCF7wPoUGBSrDGeoY8dSTBhuv21Fzw_CPEa5KeI3MOYgSN3zA1o_ZlKwjHgpSsPM3xAO_DQ; HttpOnly; Secure; SameSite=Strict; Path=/auth/refresh; Max-Age=604800"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
    delete:
      tags:
        - auth
      summary: Delete access and refresh tokens
      security:
        - cookieAuth: [ ]
      responses:
        204:
          description: Access and refresh tokens deleted successfully
          headers:
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "accessToken=; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=0"
                  - "refreshToken=; HttpOnly; Secure; SameSite=Strict; Path=/auth/refresh; Max-Age=0"
  /auth/tokens/devices:
    get:
      tags:
        - auth
      summary: Get all logged-in devices
      security:
        - cookieAuth: [ ]
      responses:
        200:
          description: Devices retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceDTO"
              example:
                - deviceId: "550e8400-e29b-41d4-a716-446655440000"
                  userAgent: "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36"
                  createdAt: "2024-01-15T10:30:00"
                  lastActiveAt: "2024-01-15T12:00:00"
                  current: true
                - deviceId: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                  userAgent: "Mozilla/5.0 (iPhone; CPU iPhone OS 17_0)"
                  createdAt: "2024-01-14T08:00:00"
                  lastActiveAt: "2024-01-14T20:00:00"
                  current: false
        401:
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags:
        - auth
      summary: Delete access and refresh tokens on all devices
      security:
        - cookieAuth: [ ]
      responses:
        204:
          description: Access and refresh tokens deleted successfully on all devices
          headers:
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "accessToken=; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=0"
                  - "refreshToken=; HttpOnly; Secure; SameSite=Strict; Path=/auth/refresh; Max-Age=0"
  /auth/tokens/devices/{deviceId}:
    delete:
      tags:
        - auth
      summary: Delete a specific device session
      security:
        - cookieAuth: [ ]
      parameters:
        - name: deviceId
          in: path
          description: Device ID to delete
          required: true
          schema:
            type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        204:
          description: Device session deleted successfully
        401:
          $ref: "#/components/responses/Unauthorized"
  /auth/tokens/refresh:
    post:
      tags:
        - auth
      summary: Refresh auth tokens
      parameters:
        - in: cookie
          name: refreshToken
          schema:
            type: string
            example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJVU0VSIl0sImlzc3VlZEF0IjoiMjAyMS0xMC0xOVQxMzowMDowMy45MDYzNDQyMDJaIiwiZXhwIjoxNjM1MjUzMjAzLCJ1c2VySWQiOjJ9.RHzh6qyGJEKYdvCuCF7wPoUGBSrDGeoY8dSTBhuv21Fzw_CPEa5KeI3MOYgSN3zA1o_ZlKwjHgpSsPM3xAO_DQ"
      responses:
        204:
          description: Auth tokens refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "accessToken=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJVU0VSIl0sImlzc3VlZEF0IjoiMjAyMS0xMC0yMFQxNDozMzo0NC43MDcyOTMzMDlaIiwiZXhwIjoxNjM0NzQxMzI0LCJ1c2VySWQiOjJ9.uXOVA1q-o2DtHmwBAzEfqEm8GLpAhXrYo0rlZ_6NFbBGILhkV74x-Iu9W2uSfSlwp1IfKPCHlR6zWVPvAbhWVw; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=9000"
                  - "refreshToken=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJVU0VSIl0sImlzc3VlZEF0IjoiMjAyMS0xMC0xOVQxMzowMDowMy45MDYzNDQyMDJaIiwiZXhwIjoxNjM1MjUzMjAzLCJ1c2VySWQiOjJ9.RHzh6qyGJEKYdvCuCF7wPoUGBSrDGeoY8dSTBhuv21Fzw_CPEa5KeI3MOYgSN3zA1o_ZlKwjHgpSsPM3xAO_DQ; HttpOnly; Secure; SameSite=Strict; Path=/auth/refresh; Max-Age=604800"
        400:
          $ref: "#/components/responses/BadRequest"
  /auth/password/forgot:
    post:
      tags:
        - auth
      summary: Send forgot password email
      requestBody:
        $ref: "#/components/requestBodies/ForgotPasswordRequest"
      responses:
        204:
          description: Forgot password email sent successfully
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
  /auth/password/reset:
    post:
      tags:
        - auth
      summary: Reset password
      requestBody:
        $ref: "#/components/requestBodies/ResetPasswordRequest"
      responses:
        204:
          description: Password reset successfully
        400:
          $ref: "#/components/responses/BadRequest"
  /auth/verification-email:
    post:
      tags:
        - auth
      summary: Send verification email
      requestBody:
        $ref: "#/components/requestBodies/VerificationEmailRequest"
      responses:
        204:
          description: Verification email sent successfully
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          $ref: "#/components/responses/Conflict"
  /auth/verify-email:
    post:
      tags:
        - auth
      summary: Verify email
      requestBody:
        $ref: "#/components/requestBodies/VerifyEmailRequest"
      responses:
        204:
          description: Email verified successfully
        400:
          $ref: "#/components/responses/BadRequest"
        409:
          $ref: "#/components/responses/Conflict"
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            timestamp: 2023-05-18T00:51:50.758738
            status: 400
            error: Bad Request
            codes:
              - API_ERROR_1
              - API_ERROR_2
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            timestamp: 2023-05-18T00:51:50.758738
            status: 401
            error: Unauthorized
            codes:
              - API_ERROR_1
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            timestamp: 2023-05-18T00:51:50.758738
            status: 403
            error: Forbidden
            codes:
              - API_ERROR_1
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            timestamp: 2023-05-18T00:51:50.758738
            status: 404
            error: Not Found
            codes:
              - API_ERROR_1
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            timestamp: 2023-05-18T00:51:50.758738
            status: 409
            error: Conflict
            codes:
              - API_ERROR_1
...
  requestBodies:
    RegisterUserRequest:
      description: Register user request
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - username
              - email
              - password
              - confirmPassword
            properties:
              username:
                type: string
              email:
                type: string
                format: email
              password:
                type: string
              confirmPassword:
                type: string
          example:
            username: John.doe
            email: john.doe@localhost
            password: qwerty123
            confirmPassword: qwerty123
    AuthTokenRequest:
      description: Auth token request
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - usernameOrEmail
              - password
            properties:
              usernameOrEmail:
                type: string
              password:
                type: string
          example:
            usernameOrEmail: user
            password: qwerty123
...
    ForgotPasswordRequest:
      description: Forgot password request
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - email
            properties:
              email:
                type: string
                format: email
          example:
            email: user@localhost
    ResetPasswordRequest:
      description: Reset password request
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - token
              - password
              - confirmPassword
            properties:
              token:
                type: string
                format: JWT
              password:
                type: string
              confirmPassword:
                type: string
          example:
            token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJVU0VSIl0sImlzc3VlZEF0IjoiMjAyMS0xMC0yMFQxNDozMzo0NC43MDcyOTMzMDlaIiwiZXhwIjoxNjM0NzQxMzI0LCJ1c2VySWQiOjJ9.uXOVA1q-o2DtHmwBAzEfqEm8GLpAhXrYo0rlZ_6NFbBGILhkV74x-Iu9W2uSfSlwp1IfKPCHlR6zWVPvAbhWVw
            password: qwerty321
            confirmPassword: qwerty321
    VerificationEmailRequest:
      description: Verification email request
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - usernameOrEmail
            properties:
              usernameOrEmail:
                type: string
          example:
            usernameOrEmail: user@localhost
    VerifyEmailRequest:
      description: Verify email request
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - token
            properties:
              token:
                type: string
                format: JWT
          example:
            token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJVU0VSIl0sImlzc3VlZEF0IjoiMjAyMS0xMC0yMFQxNDozMzo0NC43MDcyOTMzMDlaIiwiZXhwIjoxNjM0NzQxMzI0LCJ1c2VySWQiOjJ9.uXOVA1q-o2DtHmwBAzEfqEm8GLpAhXrYo0rlZ_6NFbBGILhkV74x-Iu9W2uSfSlwp1IfKPCHlR6zWVPvAbhWVw
...
  schemas:
    UserDTO:
      type: object
      required:
        - id
        - username
        - email
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
          nullable: true
        active:
          type: boolean
        lock:
          type: boolean
        createdAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleDTO"
    RoleDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          enum:
            - USER
            - ADMIN
    DeviceDTO:
      type: object
      required:
        - deviceId
        - createdAt
        - lastActiveAt
        - current
      properties:
        deviceId:
          type: string
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
        current:
          type: boolean
...
    ErrorMessage:
      type: object
      required:
        - timestamp
        - status
        - error
        - codes
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        codes:
          type: array
          items:
            type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken